a couple errors here and there

+++ LaunchpadReloaded/Buttons/CallButton.cs
@@ -15,7 +15,7 @@ public class CallButton : CustomActionButton
     public override string Name => "CALL";
     public override float Cooldown => ModdedGroupSingleton<CaptainOptions>.Instance.CaptainMeetingCooldown;
     public override float EffectDuration => 0;
-    public override int MaxUses => ModdedGroupSingleton<CaptainOptions>.Instance.CaptainMeetingCount;
+    public override int MaxUses => (int)ModdedGroupSingleton<CaptainOptions>.Instance.CaptainMeetingCount;
     public override LoadableAsset<Sprite> Sprite => LaunchpadAssets.CallButton;
 
     public override bool Enabled(RoleBehaviour role)
